Add unit test for venice-common changes
